refactor: Header 컴포넌트를 layout 폴더로 분류
Header 컴포넌트는 모든 페이지에서 전역으로 쓰일 수 있는 컴포넌트이므로 layout 폴더로 관리한다.
