+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { RecoilRoot } from 'recoil';
-import { Header } from './components/Header';
+import { Header } from './layouts/Header';
 import { ProductList } from './pages/ProductList';
 
 function App() {
