+++ src/components/Logo.tsx
@@ -6,7 +6,7 @@ export const Logo = ({ $color }: { $color: string }) => {
   return (
     <LogoWrapper>
       <LogoImage fill={$color} />
-      <LogoTitle size={'40px'} weight={'900'} $color={$color}>
+      <LogoTitle size="40px" weight="900" $color={$color}>
         SHOP
       </LogoTitle>
     </LogoWrapper>
